+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 2019-12-19 v3.1.0-SNAPSHOT
 * Updated RMB to version 29.1.5
+* Added records metadata filling for batch api
 
 ## 2019-12-17 v3.0.1
 * Fixed timeout on GET /source-storage/sourceRecords